Moved verification functions to separate file. Added verification checks to MKVFile and MKVTrack classes.
