+++ pymkv/__init__.py
@@ -1,5 +1,6 @@
 from .MKVTrack import *
 from .MKVFile import *
 from .Timestamp import *
+from .Verifications import *
 
 __version__ = '0.1.0'
